Move VtValueGet functions to libs/common #1767
