+++ libs/translator/reader/utils.cpp
@@ -33,6 +33,7 @@
 #include <vector>
 
 #include <constant_strings.h>
+#include <parameters_utils.h>
 
 #include "reader.h"
